+++ g2config/g2config.go
@@ -25,10 +25,11 @@ import (
 // ----------------------------------------------------------------------------
 
 type G2config struct {
-	GrpcClient g2pb.G2ConfigClient
-	isTrace    bool // Performance optimization
-	logger     logging.LoggingInterface
-	observers  subject.Subject
+	GrpcClient     g2pb.G2ConfigClient
+	isTrace        bool // Performance optimization
+	logger         logging.LoggingInterface
+	observerOrigin string
+	observers      subject.Subject
 }
 
 // ----------------------------------------------------------------------------
@@ -44,7 +45,7 @@ func (client *G2config) getLogger() logging.LoggingInterface {
 		options := []interface{}{
 			&logging.OptionCallerSkip{Value: 4},
 		}
-		client.logger, err = logging.NewSenzingSdkLogger(ProductId, g2configapi.IdMessages, options...)
+		client.logger, err = logging.NewSenzingSdkLogger(ComponentId, g2configapi.IdMessages, options...)
 		if err != nil {
 			panic(err)
 		}
@@ -100,7 +101,7 @@ func (client *G2config) AddDataSource(ctx context.Context, configHandle uintptr,
 				"inputJson": inputJson,
 				"return":    result,
 			}
-			notifier.Notify(ctx, client.observers, ProductId, 8001, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8001, err, details)
 		}()
 	}
 	return result, err
@@ -129,7 +130,7 @@ func (client *G2config) Close(ctx context.Context, configHandle uintptr) error {
 	if client.observers != nil {
 		go func() {
 			details := map[string]string{}
-			notifier.Notify(ctx, client.observers, ProductId, 8002, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8002, err, details)
 		}()
 	}
 	return err
@@ -163,7 +164,7 @@ func (client *G2config) Create(ctx context.Context) (uintptr, error) {
 	if client.observers != nil {
 		go func() {
 			details := map[string]string{}
-			notifier.Notify(ctx, client.observers, ProductId, 8003, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8003, err, details)
 		}()
 	}
 	return result, err
@@ -196,7 +197,7 @@ func (client *G2config) DeleteDataSource(ctx context.Context, configHandle uintp
 			details := map[string]string{
 				"inputJson": inputJson,
 			}
-			notifier.Notify(ctx, client.observers, ProductId, 8004, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8004, err, details)
 		}()
 	}
 	return err
@@ -222,10 +223,23 @@ func (client *G2config) Destroy(ctx context.Context) error {
 	if client.observers != nil {
 		go func() {
 			details := map[string]string{}
-			notifier.Notify(ctx, client.observers, ProductId, 8005, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8005, err, details)
 		}()
 	}
 	return err
+}
+
+/*
+The GetObserverOrigin method returns the "origin" value of past Observer messages.
+
+Input
+  - ctx: A context to control lifecycle.
+
+Output
+  - The value sent in the Observer's "origin" key/value pair.
+*/
+func (client *G2config) GetObserverOrigin(ctx context.Context) string {
+	return client.observerOrigin
 }
 
 /*
@@ -246,7 +260,7 @@ func (client *G2config) GetSdkId(ctx context.Context) string {
 	if client.observers != nil {
 		go func() {
 			details := map[string]string{}
-			notifier.Notify(ctx, client.observers, ProductId, 8010, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8010, err, details)
 		}()
 	}
 	return "grpc"
@@ -283,7 +297,7 @@ func (client *G2config) Init(ctx context.Context, moduleName string, iniParams s
 				"moduleName":     moduleName,
 				"verboseLogging": strconv.Itoa(verboseLogging),
 			}
-			notifier.Notify(ctx, client.observers, ProductId, 8006, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8006, err, details)
 		}()
 	}
 	return err
@@ -318,7 +332,7 @@ func (client *G2config) ListDataSources(ctx context.Context, configHandle uintpt
 	if client.observers != nil {
 		go func() {
 			details := map[string]string{}
-			notifier.Notify(ctx, client.observers, ProductId, 8007, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8007, err, details)
 		}()
 	}
 	return result, err
@@ -349,7 +363,7 @@ func (client *G2config) Load(ctx context.Context, configHandle uintptr, jsonConf
 	if client.observers != nil {
 		go func() {
 			details := map[string]string{}
-			notifier.Notify(ctx, client.observers, ProductId, 8008, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8008, err, details)
 		}()
 	}
 	return err
@@ -378,7 +392,7 @@ func (client *G2config) RegisterObserver(ctx context.Context, observer observer.
 			details := map[string]string{
 				"observerID": observer.GetObserverId(ctx),
 			}
-			notifier.Notify(ctx, client.observers, ProductId, 8011, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8011, err, details)
 		}()
 	}
 	return err
@@ -413,7 +427,7 @@ func (client *G2config) Save(ctx context.Context, configHandle uintptr) (string,
 	if client.observers != nil {
 		go func() {
 			details := map[string]string{}
-			notifier.Notify(ctx, client.observers, ProductId, 8009, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8009, err, details)
 		}()
 	}
 	return result, err
@@ -443,10 +457,21 @@ func (client *G2config) SetLogLevel(ctx context.Context, logLevelName string) er
 			details := map[string]string{
 				"logLevel": logLevelName,
 			}
-			notifier.Notify(ctx, client.observers, ProductId, 8012, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8012, err, details)
 		}()
 	}
 	return err
+}
+
+/*
+The SetObserverOrigin method sets the "origin" value in future Observer messages.
+
+Input
+  - ctx: A context to control lifecycle.
+  - origin: The value sent in the Observer's "origin" key/value pair.
+*/
+func (client *G2config) SetObserverOrigin(ctx context.Context, origin string) {
+	client.observerOrigin = origin
 }
 
 /*
@@ -471,7 +496,7 @@ func (client *G2config) UnregisterObserver(ctx context.Context, observer observe
 		details := map[string]string{
 			"observerID": observer.GetObserverId(ctx),
 		}
-		notifier.Notify(ctx, client.observers, ProductId, 8013, err, details)
+		notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8013, err, details)
 	}
 	err = client.observers.UnregisterObserver(ctx, observer)
 	if !client.observers.HasObservers(ctx) {