+++ g2config/main.go
@@ -5,4 +5,4 @@ package g2config
 // ----------------------------------------------------------------------------
 
 // Identfier of the g2config package found messages having the format "senzing-6021xxxx".
-const ProductId = 6021
+const ComponentId = 6021