+++ g2configmgr/g2configmgr.go
@@ -25,10 +25,11 @@ import (
 // ----------------------------------------------------------------------------
 
 type G2configmgr struct {
-	GrpcClient g2pb.G2ConfigMgrClient
-	isTrace    bool // Performance optimization
-	logger     logging.LoggingInterface
-	observers  subject.Subject
+	GrpcClient     g2pb.G2ConfigMgrClient
+	isTrace        bool // Performance optimization
+	logger         logging.LoggingInterface
+	observerOrigin string
+	observers      subject.Subject
 }
 
 // ----------------------------------------------------------------------------
@@ -44,7 +45,7 @@ func (client *G2configmgr) getLogger() logging.LoggingInterface {
 		options := []interface{}{
 			&logging.OptionCallerSkip{Value: 4},
 		}
-		client.logger, err = logging.NewSenzingSdkLogger(ProductId, g2configmgrapi.IdMessages, options...)
+		client.logger, err = logging.NewSenzingSdkLogger(ComponentId, g2configmgrapi.IdMessages, options...)
 		if err != nil {
 			panic(err)
 		}
@@ -97,7 +98,7 @@ func (client *G2configmgr) AddConfig(ctx context.Context, configStr string, conf
 			details := map[string]string{
 				"configComments": configComments,
 			}
-			notifier.Notify(ctx, client.observers, ProductId, 8001, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8001, err, details)
 		}()
 	}
 	return result, err
@@ -123,7 +124,7 @@ func (client *G2configmgr) Destroy(ctx context.Context) error {
 	if client.observers != nil {
 		go func() {
 			details := map[string]string{}
-			notifier.Notify(ctx, client.observers, ProductId, 8002, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8002, err, details)
 		}()
 	}
 	return err
@@ -157,7 +158,7 @@ func (client *G2configmgr) GetConfig(ctx context.Context, configID int64) (strin
 	if client.observers != nil {
 		go func() {
 			details := map[string]string{}
-			notifier.Notify(ctx, client.observers, ProductId, 8003, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8003, err, details)
 		}()
 	}
 	return result, err
@@ -188,7 +189,7 @@ func (client *G2configmgr) GetConfigList(ctx context.Context) (string, error) {
 	if client.observers != nil {
 		go func() {
 			details := map[string]string{}
-			notifier.Notify(ctx, client.observers, ProductId, 8004, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8004, err, details)
 		}()
 	}
 	return result, err
@@ -218,10 +219,23 @@ func (client *G2configmgr) GetDefaultConfigID(ctx context.Context) (int64, error
 	if client.observers != nil {
 		go func() {
 			details := map[string]string{}
-			notifier.Notify(ctx, client.observers, ProductId, 8005, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8005, err, details)
 		}()
 	}
 	return result, err
+}
+
+/*
+The GetObserverOrigin method returns the "origin" value of past Observer messages.
+
+Input
+  - ctx: A context to control lifecycle.
+
+Output
+  - The value sent in the Observer's "origin" key/value pair.
+*/
+func (client *G2configmgr) GetObserverOrigin(ctx context.Context) string {
+	return client.observerOrigin
 }
 
 /*
@@ -242,7 +256,7 @@ func (client *G2configmgr) GetSdkId(ctx context.Context) string {
 	if client.observers != nil {
 		go func() {
 			details := map[string]string{}
-			notifier.Notify(ctx, client.observers, ProductId, 8010, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8010, err, details)
 		}()
 	}
 	return "grpc"
@@ -279,7 +293,7 @@ func (client *G2configmgr) Init(ctx context.Context, moduleName string, iniParam
 				"moduleName":     moduleName,
 				"verboseLogging": strconv.Itoa(verboseLogging),
 			}
-			notifier.Notify(ctx, client.observers, ProductId, 8006, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8006, err, details)
 		}()
 	}
 	return err
@@ -308,7 +322,7 @@ func (client *G2configmgr) RegisterObserver(ctx context.Context, observer observ
 			details := map[string]string{
 				"observerID": observer.GetObserverId(ctx),
 			}
-			notifier.Notify(ctx, client.observers, ProductId, 8010, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8010, err, details)
 		}()
 	}
 	return err
@@ -343,7 +357,7 @@ func (client *G2configmgr) ReplaceDefaultConfigID(ctx context.Context, oldConfig
 			details := map[string]string{
 				"newConfigID": strconv.FormatInt(newConfigID, 10),
 			}
-			notifier.Notify(ctx, client.observers, ProductId, 8007, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8007, err, details)
 		}()
 	}
 	return err
@@ -374,7 +388,7 @@ func (client *G2configmgr) SetDefaultConfigID(ctx context.Context, configID int6
 			details := map[string]string{
 				"configID": strconv.FormatInt(configID, 10),
 			}
-			notifier.Notify(ctx, client.observers, ProductId, 8008, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8008, err, details)
 		}()
 	}
 	return err
@@ -404,10 +418,21 @@ func (client *G2configmgr) SetLogLevel(ctx context.Context, logLevelName string)
 			details := map[string]string{
 				"logLevel": logLevelName,
 			}
-			notifier.Notify(ctx, client.observers, ProductId, 8011, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8011, err, details)
 		}()
 	}
 	return err
+}
+
+/*
+The SetObserverOrigin method sets the "origin" value in future Observer messages.
+
+Input
+  - ctx: A context to control lifecycle.
+  - origin: The value sent in the Observer's "origin" key/value pair.
+*/
+func (client *G2configmgr) SetObserverOrigin(ctx context.Context, origin string) {
+	client.observerOrigin = origin
 }
 
 /*
@@ -432,7 +457,7 @@ func (client *G2configmgr) UnregisterObserver(ctx context.Context, observer obse
 		details := map[string]string{
 			"observerID": observer.GetObserverId(ctx),
 		}
-		notifier.Notify(ctx, client.observers, ProductId, 8012, err, details)
+		notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8012, err, details)
 	}
 	err = client.observers.UnregisterObserver(ctx, observer)
 	if !client.observers.HasObservers(ctx) {