+++ g2configmgr/main.go
@@ -5,4 +5,4 @@ package g2configmgr
 // ----------------------------------------------------------------------------
 
 // Identfier of the g2configmgr package found messages having the format "senzing-6022xxxx".
-const ProductId = 6022
+const ComponentId = 6022