+++ g2diagnostic/g2diagnostic_test.go
@@ -249,7 +249,7 @@ func setup() error {
 	options := []interface{}{
 		&logging.OptionCallerSkip{Value: 4},
 	}
-	localLogger, err = logging.NewSenzingSdkLogger(ProductId, g2diagnosticapi.IdMessages, options...)
+	localLogger, err = logging.NewSenzingSdkLogger(ComponentId, g2diagnosticapi.IdMessages, options...)
 	if err != nil {
 		return createError(5901, err)
 	}
@@ -286,6 +286,22 @@ func teardown() error {
 // ----------------------------------------------------------------------------
 // Test interface functions
 // ----------------------------------------------------------------------------
+
+func TestG2diagnostic_SetObserverOrigin(test *testing.T) {
+	ctx := context.TODO()
+	g2diagnostic := getTestObject(ctx, test)
+	origin := "Machine: nn; Task: UnitTest"
+	g2diagnostic.SetObserverOrigin(ctx, origin)
+}
+
+func TestG2diagnostic_GetObserverOrigin(test *testing.T) {
+	ctx := context.TODO()
+	g2diagnostic := getTestObject(ctx, test)
+	origin := "Machine: nn; Task: UnitTest"
+	g2diagnostic.SetObserverOrigin(ctx, origin)
+	actual := g2diagnostic.GetObserverOrigin(ctx)
+	assert.Equal(test, origin, actual)
+}
 
 func TestG2diagnostic_CheckDBPerf(test *testing.T) {
 	ctx := context.TODO()
@@ -493,6 +509,26 @@ func TestG2diagnostic_Destroy(test *testing.T) {
 // ----------------------------------------------------------------------------
 // Examples for godoc documentation
 // ----------------------------------------------------------------------------
+
+func ExampleG2diagnostic_SetObserverOrigin() {
+	// For more information, visit https://github.com/Senzing/g2-sdk-go-grpc/blob/main/g2diagnostic/g2diagnostic_test.go
+	ctx := context.TODO()
+	g2diagnostic := getG2Diagnostic(ctx)
+	origin := "Machine: nn; Task: UnitTest"
+	g2diagnostic.SetObserverOrigin(ctx, origin)
+	// Output:
+}
+
+func ExampleG2diagnostic_GetObserverOrigin() {
+	// For more information, visit https://github.com/Senzing/g2-sdk-go-grpc/blob/main/g2config/g2diagnostic_test.go
+	ctx := context.TODO()
+	g2diagnostic := getG2Diagnostic(ctx)
+	origin := "Machine: nn; Task: UnitTest"
+	g2diagnostic.SetObserverOrigin(ctx, origin)
+	result := g2diagnostic.GetObserverOrigin(ctx)
+	fmt.Println(result)
+	// Output: Machine: nn; Task: UnitTest
+}
 
 func ExampleG2diagnostic_CheckDBPerf() {
 	// For more information, visit https://github.com/Senzing/g2-sdk-go-grpc/blob/main/g2diagnostic/g2diagnostic_test.go