+++ g2diagnostic/main.go
@@ -5,4 +5,4 @@ package g2diagnostic
 // ----------------------------------------------------------------------------
 
 // Identfier of the g2diagnostic package found messages having the format "senzing-6023xxxx".
-const ProductId = 6023
+const ComponentId = 6023