+++ g2engine/main.go
@@ -5,4 +5,4 @@ package g2engine
 // ----------------------------------------------------------------------------
 
 // Identfier of the g2engine package found messages having the format "senzing-6024xxxx".
-const ProductId = 6024
+const ComponentId = 6024