+++ g2product/main.go
@@ -5,4 +5,4 @@ package g2product
 // ----------------------------------------------------------------------------
 
 // Identfier of the g2product package found messages having the format "senzing-6026xxxx".
-const ProductId = 6026
+const ComponentId = 6026